ctf: Commit the generated ANTLR files to the source tree
This means users who just pulled the git tree won't need to run
a Maven build manually to generate them. Which in turn, completely
removes the need for installing m2e or such in their Eclipse.

They are now only built when the "ctf-grammar" profile is enabled.
Rebuilding always produces a diff: the first line of the generated
files contains a timestamp, and there is some randomness in the
generated comments. So it's better to only do it on-demand.

Also added a copyright header to these files, since they didn't
seem to have one.

Change-Id: I0fcf4cb32f111c7eecde3ab641205fb1d526fb28
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>