lttng: Support for existing live sessions
With this change, it is now possible to create a live session from the
command line and connect to it using the Control view as a live session.
We use the live timer interval as an indicator of whether or not a
session is live. The information about the host and the port is not
available with 'lttng list' but only local host is supported for now
anyway (127.0.0.1). For the port, we try use the default (5344). For
now, if the default port is incorrect, an error message "The connection
could not be established" is shown. As a future enhancement, there
should be a dialog to configure the relayd connection independently from
the session.

Change-Id: I9a56426ca011441ed5ebb3d8f936ababf6cdbce1
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>