tmf.ctf: Rename CtfIteratorManager classes
The goal of this patch series is to clean up the whole CtfIteratorManager
mess. Similarly to the TmfEventTypeManager, there is no point have all iterators
registered to a central manager, they should be encapsulated into the trace
they belong to.

This first patch simply renames existing constructs to better represent what
they do, which should help future refactoring:

CtfIteratorManager -> CtfIteratorManagerManager .It is a "meta-manager" that
manages all the existing, now-called CtfIteratorManager's. Silly name, but
don't worry it's temporary ;)

CtfTraceManager -> CtfIteratorManager. This is the object that manages all
the iterators of one trace. It should eventually be moved to inside the trace.

Change-Id: Ie5483844856610907d3e332307eb0844209a93ef
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>