tmf: Introduce the core concept of event aspects
An "event aspect" is a piece of information that can be extracted,
directly or indirectly, from a trace event. Simple examples could be
timestamp or event fields. But it could also be something indirect,
like a state system request at the timestamp of the given event.

This patch introduces the base ITmfEventAspect interface, some basic
implementations, and simplifies the TmfEventTableColumns to use aspects
instead, since part of it is now redundant.

Fixes bug #451411.

Change-Id: Ib045439d6084a6e6c7c276f1eddccfb13fe01be0
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>