TMF: Add protected method to AbstracTimeGraphView to fetch all entries at once
This gives another option for views to set the events for entries. Right now,
it is only possible to get events for one entry at a time. It supposes that
each entry is independent from the others. If it is not the case, this new
method can be overridden and the view can fetch the events for many entries
at the same time.

Change-Id: I057e72f53094520eb6751e986a8bc037daed6ae5
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>