tmf: Remove the ITmfEventTableColumns extension point
Add a getCriteria() method in ITmfTrace, which allows trace types
to define which criteria can be relevant for their own events.

This becomes redundant with the ITmfEventTableColumn extension point:
now that we build columns directly from criteria, instead of reading
extension point data the TmfEventsEditor can simply call
trace.getCriteria() instead. The trace type is already defined in a
core extension point, so we'll pass the criteria information through
that now.

The combination of careful equals() methods in criteria, and the use
of a Set in the TmfEventsEditor, will allow "merging" criteria/columns
together, like was already the case.

Change-Id: Ic5b6126a50b1b1fdb4b2d5d23eb6a47a57555594
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>