tmf.analysis: Restore the "getCPU" for data-driven analyses
XML data-driven analyses used to use ITmfEvent.getSource() to get
the event's CPU. This was misleading, because trace types were free
to define "source" as something other than CPUs.

Now instead use a CpuAspect to get the CPU of an event. This
should restore support for getting the CPU of CTF traces, as it
did before.

Also added a new method to TmfTraceUtils to get the aspects matching
a specific class. The analysis modules will make use of this to get
the "cpu" aspects.

Change-Id: I88437b4a419e81aef36f2aa5298ff54f87242b23
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>