tmf: Make Pair utility class non-null and immutable
Nullable generic types summon all sorts of hells with Java 8
type annotations. Making this class immutable and non-null
makes it much easier to handle.

This required tweaking EventIterator to not use a Pair anymore, because
it used null values. Instead, we provide two separate methods to get
the first and second parts of a split.

Change-Id: I58007eaa488db337d0f34698e3144f232ee3c784
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>