TMF: Make the ITmfEventAspect#resolve nullable
A null value has the advantage to really show an aspect is not available for
a given event instead of having to find of dummy empty objects that we need to
compare with in most cases.

With the new common NonNullUtils class, it is now easy to convert a null value
to something useful, like a nonNull String, so it does not add useless null
checks to the code.

Change-Id: I9f41ec6522b5132244c630ce9a6b174599802954
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>