TMF: Add utility method to return the result of an aspect for an event
This method avoids having to iterate through an event aspect iterable and
breaking as soon as it resolves to nonNull for a given event. The result of
the aspect's resolve(ITmfEvent) is returned. It can be null, but if it is
not, it can be safely cast to the aspect's return type.

Change-Id: I83cd2e56ad30c77608d9b184ede6d9df3240393a
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>