tmf: Replace TmfFilterEventTypeNode with TmfFilterTraceTypeNode
The filtering on event type is replaced with filtering on trace type.
This is done because in the upcoming change to replace field-based
filtering with aspect-based filtering, the aspect list is dependent on
the trace type and not on the event type. Furthermore filtering on event
type will still be possible since an 'Event type' aspect is available.

With this patch, the narrowing of field list based on event type for
child nodes of an EVENTTYPE node is no longer supported. In the
upcoming patch the aspect list will be narrowed based on trace type for
child nodes of a TRACETYPE node.

Change-Id: Ie911e9c63ecf7e219540e038fd82e70415e2e0ee
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>