TMF: Introduce a fast linear timestamp transform
The fast timestamp transform uses standard integer arithmetic to compute the
the transform, instead of BigDecimal, yet producing monotonic timestamp within
3ns of the original function.

Test that the fast transform yields about the same results as the original
function, either forward or backward, and also verify that the cache is
effectively used.

The benchmark results are:

Transform   | time (ms)
-----------------------
original    |     17310
fast        |       256

This benchmark suggests that the fast transform is about 67 times faster than
the original transform.

Change-Id: Ifdf8e23b3e042bf1f2b0454e64a4ab4c47d408d4
Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>