TMF: Introduce a fast linear timestamp transform
The fast timestamp transform uses standard integer arithmetic to compute the
the transform, instead of BigDecimal, yet producing monotonic timestamp within
3ns of the original function.

Test that the fast transform yields about the same results as the original
function, either forward or backward, and also verify that the cache is
effectively used.

The benchmark results are:

Transform   | time (ms)
-----------------------
original    |     17800
fast        |       311

This benchmark suggests that the fast transform is about 57 times faster than
the original transform.

This patch adds a dependency on com.google.common.hash to use hash function
from guava instead of reinventing the wheel.

Change-Id: Ifdf8e23b3e042bf1f2b0454e64a4ab4c47d408d4
Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>