ss: Add a "checkpoint" attribute to AbstractTmfStateProvider
This attribute will be immediately useful only for the partial state
history, but it could be used by any backend (or even view) for indexing
or checkpointing of any kind.

The only other solution to pass event information to a backend would be
by doing some weird hacks with event requests, which doesn't feel as
robust as the solution here.

Change-Id: I0c135badb7f847bc760d5109fe9368122e1c7f55
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>