ctf: Remove getMaximumSize() from IDeclaration
A maximum size can be known at declaration level for types like
integers (and compound types using subtypesgetMaximumSize()), but
not for things like Strings, so it does not make much sense to have
this in the general interface.

It was also only used in exactly one place : a check in the initialization
of sequences. Instead, replace this check with a completely arbitrary
implementation limit.

This will allow actually reading sequences of strings, which would
miserably fail before (bug #457951).

Change-Id: I14c680710f4d4fa691484494957082db7d17035c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>