ctf: Avoid direct dependency from the lttng plugins to the ctf.core one
The tmf.ctf plugin should encapsulate ctf.core's API completely, so that
lttng (and other eventual users) do not need to depend on ctf.core directly.

Move CtfIterator to be internal, and do not expose it in CtfTmfTrace's public
API. Instead, return ITmfContext's, which can be cast to CtfIterator for
things like tests that really the a CtfIterator reference.

Change-Id: Id7c74c284af98001ca0d0a63648e0db6219c2d5c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>