ctf: Avoid direct dependency from the lttng plugins to the ctf.core one
The tmf.ctf plugin should encapsulate ctf.core's API completely, so that
lttng (and other eventual users) do not need to depend on ctf.core directly.

The biggest hurdle for this is the poorly-designed CtfIterator: instead of
extending the TMF component and *containing* the CTF one, like we do in
every single other extension, it does it the other way around. Its
constructor then requires getting the CTFTrace from a CtfTmfTrace, which
we cannot do if we do not want to expose the CTFTrace publicly.

The workaround used was to provide a utility class that can retrieve the
field via reflection. It works, but is ugly. Hopefully it is temporary.

Change-Id: Id7c74c284af98001ca0d0a63648e0db6219c2d5c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>