btf: Fix BTF event property source
The BTF UI plug-in is now started through org.eclipse.core.runtime.adapters
extension, since no class would trigger the plug-in start. This allows
the BTF event adapter factory to be registered. To force loading the plug-in,
we let BtfEvent load the adapter instead of simply getting it which in turn
creates the factory and loads the plug-in.

The BTF event property source is updated to display all BTF event
aspects into their own property.

Change-Id: I37fe006cece41f9b470da714314cef932ab7e9a8
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>