tmf: Don't import uneeded entries in FunctionNameMapper
Lines from a 'nm' output file that represent non-static symbols
(so for which we don't have adresses) would create an entry in
the function map with an empty string as address.

This is because String.split(" ") would split the series of spaces
at the beginning of the line into many sub-strings. Simply doing
String.trim() first avoids that.

In practice it did not matter much because no entry in the view would
use the empty string key. Still, it's cleaner to fix it.

Change-Id: I4a3ac452aedd8d8cf4ff85db982ccd2c9fdf5693
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>