tmf: Fix time graph bounds reset on refresh and ignoring child entries
Calling refresh() or setInput() should not reset the bounds if they have
been set to specific values by the user.

When determining bounds based on the input time graph entries, only the
root entries were considered. All roots and their recursive children
should be considered.

The value SWT.DEFAULT can now be used as a time bound to indicate that
the bound should be set according to the input time graph entries. For
any other value the bound will be fixed and unaffected by the input.

This setting can be applied independently for each bound.

The distinction between user-specified and actual bounds is made clearer
in the API Javadoc.

Change-Id: Ia79d35d41ed08cf7475807f9d352b1931196e5d5
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>