ss: Bug 454057: Support slash character in attribute names
Change the return type of getFullAttributePath() to return a String
array instead of a String.

Add StateSystemUtils.pathToString(). It converts a full path array into
a path string, and escapes '/' and '\' with a preceding '\' in
attribute names.

Add StateSystemUtils.stringToPath(). It converts a path string into a
full path array, and unescapes '/' and '\' in attribute names.

The converted path string is written as-is in the attribute tree file
store, and converted back when read from file store.

The path string is displayed in the State System Explorer view and
matches what would be expected to be entered by a end-user to
unambiguously specify the attribute using a slash-separated string.

Change-Id: Ia655fe6153665283fce6bb34ee56d37035cb5d23
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>