ss: Bug 454057: Support slash character in attribute names
The full attribute path now escapes '/' and '\' with a preceding '\' in
attribute names. The path is now unique and unambiguous.

The path is written as-is in the attribute tree file store.

The path is displayed in the State System Explorer view and matches what
would be expected to be entered by a end-user to unambiguously specify
the attribute using a slash-separated path.

Change-Id: Ia655fe6153665283fce6bb34ee56d37035cb5d23
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>