os.linux: Add a new WAIT_UNKNOWN state
The WAIT state resulting from the statedump events should be
handled slightly differently from, for example, the WAIT state
of a process being scheduled out at a sched_switch.

In the first case, we do not have any information in the trace
about *what* the process is actually waiting on. So it would
be interesting to differentiate it from a real WAIT_BLOCKED,
for which the user can follow the chain of dependencies.

With this patch, the statedump events set the processes to a new
WAIT_UNKNOWN state (and only if there is no prior status set),
different from the WAIT_BLOCKED used for active processes.

This should help a bit with the "everything-is-yellow" syndrome
that plagues your average CFV.

Thanks to Julien Desfossez for this suggestion.

Change-Id: I7d7bb1bfc1d314c822df5ee04bd8b33c6a0c7d36
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>