tmf: Fix event table item default colors and foreground positions
The default system colors should be used whenever the default color
setting from the color setting manager is used. The current theme's
system colors are not necessarily equal to COLOR_LIST_BACKGROUND and
COLOR_LIST_FOREGROUND.

The table item foreground should also be set to null in order to use the
default system color so that it can match the current theme.

The background needs to be drawn in the custom EraseItem listener if the
item background is different from the table's background (set to the
default system color as it is never changed), not when it is different
from COLOR_LIST_BACKGROUND (which may not be the default system color in
the current theme).

The text bounds and image bounds adjustments in the custom PaintItem
listener need to be set differently for Windows and Linux.

Change-Id: I06e6ed48e866ce7831ea3ad630d120a8cf228973
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>