ss: Forbid unescaped protected characters in attribute names
Add an explicit list of "protected" characters that should not be
present in attribute names, unless they are escaped by a backslash.

The first one is '/', which is used to define different levels in
the tree. See bug #454057.

I've also added ':' to the list, even though it's not really used
for anything right now, it was suggested at some point for "symlink
attributes", so we can reserve it right away. It can also confirm
that the list mechanism actually works.

Change-Id: I1b65e67f7fe8c70518fdb1b8ab5f9a9fae1798c4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>