swtbot: Fix SWTBotUtils instability
Tracing project elements that have a label with a count suffix (e.g.
trace folders) can have the count updated in the middle of a test by a
resource change event. Methods to get those tree items should be able to
handle this atomically.

Methods that need a project tree item now wait until it is available.

Change-Id: I4c2540fcc8a947ca4f61de66372968b11b10df7c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>