swtbot: Fix TestTraceOffsetting
The keystroke for enter key is fixed.

The test waits for project element to exist before opening trace.

The test waits for the trace editor to be opened or closed.

The old event table is disposed after reopening the trace and cannot be
reused. The new trace editor and its event table must be used.

The test now checks correctly that the timestamp is actually offset.

Change-Id: I069b406698207f2ca7d08f4ca87dcf0d45a4385e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>