Base code for Container Analysis
Contains several new classes as well as a modification of LttngEventLayout

LttngEventLayout modifications :
The LTTng tracer was modified and fields for namespace/containers analysis were added for the following events :
sched_process_fork, lttng_statedump_process_state
Therefore, LttngEventLayout was modified to be able to get those new fields.
The modification to the LTTng trace is pending, but a patch is available here :
https://github.com/Selora/lttng-modules/tree/patch

The rest of this commit only contains new classes for the Container Analysis.
To run the analysis on Trace Compass, one need to import the containerView.xml file.
Several traces with different use-cases are available there :
https://github.com/Selora/lttng_container_traces

To test the new analysis, a complete testbench will follow shortly

Signed-off-by: Sebastien Lorrain <sebastien.lorrain01@gmail.com>
Change-Id: Idf45b34c349faf4f26c564faa86a81445379b329