Base code for profile dialog window Modified ILttngControlService to add a loadSession command, also implemented the methods in the appropriate children classes (LTTngControlService and NullControlService).
Added ProfileDialog and ProfileHandler, as well as the appropriate modifications to the plugin.xml to allow the execution of the handler as well as the call to the Handler. As for now, pre-selected profiles must be in org.eclipse.tracecompass.internal.lttng2.control.ui/resources/folders.txt.
Appropriate .icons, strings and whatnot are externalized and at the appropriate location > messages.properties.

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@gmail.com>
Change-Id: Id312708b1907f3dd36fab65da7d96d5f917404a2