tmf remote: Decouple connection creation from JSchConnection
The o.e.remote API used to not provide an API to store username,
host and port programatically without using internal APIs of
the JSchConnection implementation. Now, o.e.remote API has been updated
to provide a generic API to set this values independent of the
actual implementation (see Bug on PTP 463468).

This patch is to update TMF remote to use the generic API.

Change-Id: Ib7ddd64d64f000e9c9e3de99e9f40afcbe10df8e
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>