ss: Bug 454057: Use serialization to write the list of attributes to disk
Simplify the writing of the Attribute Tree to disk by using the
fact that ArrayList is Serializable.

This makes the file format less portable, but the previous method
used String.getBytes(), which is also JVM-dependant. A proper
inter-operable (and more efficient) format could be implemented
at some point once the need arises.

Change-Id: I8622b0975852ed3e022a210dbd720eb4d0fe90f4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>