tmf: Remove getType() from ITmfEvent
In the great great majority of the cases, getType() was called solely
to then do getName() on the result. Now that ITmfEvent has a direct
getName(), there is little need to expose the ITmfEventType object.
The String returned by getName() plays the role of the event type now.

It is still being used in TmfEvent construction, but this becomes
an implementation detail specific to TmfEvent (not *I*TmfEvent).

This ensures all callers use ITmfEvent#getName() directly, which
always has equal or better performance.

Change-Id: I2c7b2b7ddfcaa3eef2fba719ee4a44abaf94cd59
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>