CTF: Enhance CTF validation for invalid CTF traces (Bug 464329)
The CTF validation has been updated to do first a weak validation
to determine if the directory is actually a CTF trace. After that
a full validation is done. If the full validation fails and the
weak validation was successful a TraceValidationStatus with
severity warning and confidence of 1 is return. 

The automatic trace type detection has been updated so that the
special case above is acounted for. 

With these updates the trace is imported as CTF trace however, the
use is notified that with a proper error message in case the full
validation failed.

Change-Id: Ia9b798b34c7af122d3b7684e91b114b38b656cef
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>